requirements.txt
asgiref==3.2.10
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.1
django-admin-thumbnails==0.2.6
idna==3.3
Pillow==9.1.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
